+++ NEWS.md
@@ -1,6 +1,6 @@
 ## Version 1.0.2
 
-This is a minor release, resolving some problems with `plot_discrim()` and adding functionality
+This is a minor release, resolving some problems with `plot_discrim()` and adding considerable functionality
 
 * Fixed buglet in `plot_discrim()` where the variables in a formula needed to be reversed.
 * `plot_discrim()` now gets the `data` from the object, if not supplied.
@@ -11,6 +11,7 @@ This is a minor release, resolving some problems with `plot_discrim()` and addin
 * Flesh out package description in README.
 * Default for `posterior` in `predict_discrim()` becomes `FALSE`, as rarely needed
 * `plot_discrim()` now allows plotting the results in discriminant space, e.g., `LD2 ~ LD1`
+* It also generates nice labels for the discriminant dimensions
 
 ## Version 1.0.0