+++ NEWS.md
@@ -1,4 +1,4 @@
-## Version 1.0.2
+## Version 1.1.0
 
 This is a minor release, resolving some problems with `plot_discrim()` and adding considerable functionality
 
@@ -13,6 +13,7 @@ This is a minor release, resolving some problems with `plot_discrim()` and addin
 * `plot_discrim()` now allows plotting the results in discriminant space, e.g., `LD2 ~ LD1`
 * It also generates nice labels for the discriminant dimensions
 * Added `scores.lda()` to extract discriminant scores
+* Fixed problems in reflect()
 
 ## Version 1.0.0